Added MaxDrawableIndices variable to the VertexBuffer class to ensure that the max size remains constant.
If the maximum size is not known a priori, then CalculateDynamicMaxSize can be called which uses the old method
of dynamically calculating the MaxSize based on the number of DrawableIndices.

This also fixes a bug where the MaxSize of a vertex buffer would continuously increase over the liftime
of the program, resulting in very large slowdowns.

Additionally, the BufferDataSpecification structure now contains a SizeInBytes field
which is used to make the MaxSize calculation easier.